Move password resolving to library component
